Create device-specific config
I wanted this repo to contain a central config that I could check into
git and download on all my machines without worrying about maintaining
separate configurations. As such, I separated out the configuration
into parts that are specific to my laptop and desktop, and parts that
are common to both. The device-specific configuration is chosen with a
file, select-config.lisp, which is not checked into git and only
loaded if it exists. This file should set the *selected-config*
variable to the desired machine-specific config, which will then be
loaded later.